+++ Samples~/UIToolkit/MyFirstDataBoundView/DynamicMaterialColor.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+public class DynamicMaterialColor : MonoBehaviour
+{
+    public Material material;
+    public Color Color
+    {
+        set
+        {
+            material.color = value;
+        }
+    }
+}